Update for inline party links
Previously the Party information in the Membership table was plain text.
However, the box of party info wasn't quite good enough to be able to
compare against, as it used short names rather than long, and the
Wikipedia article tiitles didn't necessarily always completely agree
with the long names either.

This has now been changed so that the first time a party is named, it
is a link.